+++ src/selection.h
@@ -13,7 +13,7 @@ typedef struct {
 void init_selection(const View *v, SelectionInfo *info);
 long prepare_selection(View *v);
 char *view_get_selection(View *v, long *size);
-int get_nr_selected_lines(const SelectionInfo *info);
-int get_nr_selected_chars(const SelectionInfo *info);
+size_t get_nr_selected_lines(const SelectionInfo *info);
+size_t get_nr_selected_chars(const SelectionInfo *info);
 
 #endif
